major revision
* Folder structure is revised.
* Documentation is revised and shortened.
* Missing `new` methods are added.
* Examples are updated.
