+++ src/dary/daryheap.rs
@@ -80,6 +80,26 @@ where
     N: Clone,
     K: PartialOrd + Clone,
 {
+    /// Creates a new empty d-ary heap.
+    ///
+    ///  # Examples
+    ///
+    /// ```
+    /// use orx_priority_queue::*;
+    ///
+    /// let mut heap = BinaryHeap::new();
+    ///
+    /// heap.push('a', 4);
+    /// heap.push('b', 42);
+    ///
+    /// assert_eq!(Some('a'), heap.pop_node());
+    /// assert_eq!(Some('b'), heap.pop_node());
+    /// assert!(heap.is_empty());
+    /// ```
+    pub fn new() -> Self {
+        Self::default()
+    }
+
     /// Creates a new d-ary heap with the given initial `capacity` on the number of nodes to simultaneously exist on the heap.
     ///
     /// # Examples
@@ -97,6 +117,7 @@ where
             heap: Heap::new(Some(capacity), HeapPositionsNone),
         }
     }
+
     /// Returns the 'd' of the d-ary heap.
     /// In other words, it represents the maximum number of children that each node on the heap can have.
     pub const fn d() -> usize {
@@ -141,32 +162,40 @@ where
     fn len(&self) -> usize {
         self.heap.len()
     }
+
     #[inline(always)]
     fn capacity(&self) -> usize {
         self.heap.capacity()
     }
+
     fn peek(&self) -> Option<&(N, K)> {
         self.heap.peek()
     }
+
     fn clear(&mut self) {
         self.heap.clear()
     }
+
     #[inline(always)]
     fn pop(&mut self) -> Option<(N, K)> {
         self.heap.pop()
     }
+
     #[inline(always)]
     fn pop_node(&mut self) -> Option<N> {
         self.heap.pop_node()
     }
+
     #[inline(always)]
     fn pop_key(&mut self) -> Option<K> {
         self.heap.pop_key()
     }
+
     #[inline(always)]
     fn push(&mut self, node: N, key: K) {
         self.heap.push(node, key)
     }
+
     #[inline(always)]
     fn push_then_pop(&mut self, node: N, key: K) -> (N, K) {
         self.heap.push_then_pop(node, key)
