+++ src/dary/daryheap_map.rs
@@ -1,50 +1,28 @@
 use super::heap::Heap;
 use crate::{
-    positions::has_index::HeapPositionsHasIndex, HasIndex, PriorityQueue, PriorityQueueDecKey,
-    ResUpdateKey,
+    positions::map::{HeapPositionsMap, Index},
+    PriorityQueue, PriorityQueueDecKey, ResUpdateKey,
 };
 
-/// Type alias for `DaryHeapOfIndices<N, K, 2>`; see [`DaryHeapOfIndices`] for details.
-pub type BinaryHeapOfIndices<N, K> = DaryHeapOfIndices<N, K, 2>;
-/// Type alias for `DaryHeapOfIndices<N, K, 4>`; see [`DaryHeapOfIndices`] for details.
-pub type QuarternaryHeapOfIndices<N, K> = DaryHeapOfIndices<N, K, 4>;
+/// Type alias for `DaryHeapWithMap<N, K, 2>`; see [`DaryHeapWithMap`] for details.
+pub type BinaryHeapWithMap<N, K> = DaryHeapWithMap<N, K, 2>;
+/// Type alias for `DaryHeapWithMap<N, K, 4>`; see [`DaryHeapWithMap`] for details.
+pub type QuarternaryHeapWithMap<N, K> = DaryHeapWithMap<N, K, 4>;
 
 /// A d-ary heap which implements both `PriorityQueue` and `PriorityQueueDecKey`.
 ///
 /// See [`PriorityQueueDecKey`] for additional functionalities.
 ///
-/// `DaryHeapOfIndices` achieves the additional features by making use of a fixed size position
-/// array which allows to track the position of nodes on the heap.
-///
-/// It has the limitation that the nodes must implement [`HasIndex`].
-/// This trait has a single simple method `fn index(&self) -> usize` which acts as a unique identifier
-/// of the actual underlying node which is coming from a closed set.
-///
-/// Consider for instance the usage of the heap as the priority queue of Dijkstra's shortest path algorithm.
-/// The nodes are actual nodes of the graph which is a closed set and can be identified by node indices from
-/// zero to `N-1`, where `N` is the number of nodes. This heap fits very well such mathematical algorithms
-/// due to the following:
-/// * using a fixed size array could be considered as a fast `HashMap`.
-/// * we often reuse such heaps many times to solve many problems on the same network,
-/// compensating for the allocation of the positions array once.
-/// * further, compared to a basic priority queue (or to `std::collections::BinaryHeap`),
-/// it reduces the space complexity of the Dijkstra's
-/// algorithm from *O(N^2)* to *O(N)* by enabling the `decrease_key` operation.
-///
-/// However, for situations where
-/// * the number of nodes entering the queue is very sparse compared to the size of the set of nodes, or
-/// * it is not convenient to index the sets,
-///
-/// `DaryHeapWithMap` provides a more flexible approach.
+/// `DaryHeapWithMap` achieves the additional features by making use of a map of nodes to positions on the heap.
 ///
 /// # Flexibility (`DaryHeapWithMap`) vs Performance (`DaryHeapOfIndices`)
 ///
-/// `DaryHeapWithMap` (hence its variants such as `BinaryHeapWithMap`) does not require to know
+/// [`DaryHeapWithMap`] (hence its variants such as [`BinaryHeapWithMap`]) does not require to know
 /// the absolute size of the closed set.
 /// Furthermore, the node type needs to implement `Hash + Eq` rather than `HasIndex` trait defined in this crate.
 /// Due to these, `DaryHeapWithMap` might be considered as the more flexible [`PriorityQueueDecKey`] variant.
 ///
-/// On the other hand, [`DaryHeapOfIndices`] (hence its variants such as [`BinaryHeapOfIndices`]),
+/// On the other hand, `DaryHeapOfIndices` (hence its variants such as `BinaryHeapOfIndices`),
 /// provides significantly faster accesses to positions of nodes on the heap.
 /// This is important for [`PriorityQueueDecKey`] operations such as `decrease_key` or `contains`.
 /// Furthermore, in many algorithms such as certain network algorithms where nodes enter and exit the queue,
@@ -82,12 +60,16 @@ pub type QuarternaryHeapOfIndices<N, K> = DaryHeapOfIndices<N, K, 4>;
 ///     assert!(pq.is_empty());
 /// }
 ///
-/// // d-hap heap using id's to locate existing nodes (although decrease-key is not used here)
-/// test_priority_queue(DaryHeapOfIndices::<_, _, 4>::with_index_bound(32));
+/// // d-ary heap using a hash map to locate existing nodes (although decrease-key is not used here)
+/// test_priority_queue(DaryHeapWithMap::<_, _, 3>::default());
+/// test_priority_queue(DaryHeapWithMap::<_, _, 4>::with_capacity(16));
 /// // using type aliases to simplify signatures
-/// test_priority_queue(BinaryHeapOfIndices::with_index_bound(16));
-/// test_priority_queue(QuarternaryHeapOfIndices::with_index_bound(16));
-/// test_priority_queue(QuarternaryHeapOfIndices::with_index_bound(16));
+/// test_priority_queue(BinaryHeapWithMap::default());
+/// test_priority_queue(BinaryHeapWithMap::with_capacity(16));
+/// test_priority_queue(QuarternaryHeapWithMap::default());
+/// test_priority_queue(QuarternaryHeapWithMap::with_capacity(16));
+/// test_priority_queue(QuarternaryHeapWithMap::default());
+/// test_priority_queue(QuarternaryHeapWithMap::with_capacity(16));
 /// ```
 ///
 /// ## Heap as a `PriorityQueueDecKey`
@@ -126,70 +108,68 @@ pub type QuarternaryHeapOfIndices<N, K> = DaryHeapOfIndices<N, K, 4>;
 ///
 ///     assert!(pq.is_empty());
 /// }
-/// // d-ary heap using id's to locate existing nodes
-/// test_priority_queue_deckey(DaryHeapOfIndices::<_, _, 3>::with_index_bound(32));
+/// // d-ary heap using a hash map to locate existing nodes
+/// test_priority_queue_deckey(DaryHeapWithMap::<_, _, 3>::default());
+/// test_priority_queue_deckey(DaryHeapWithMap::<_, _, 4>::with_capacity(16));
 /// // using type aliases to simplify signatures
-/// test_priority_queue_deckey(BinaryHeapOfIndices::with_index_bound(16));
-/// test_priority_queue_deckey(QuarternaryHeapOfIndices::with_index_bound(16));
-/// test_priority_queue_deckey(QuarternaryHeapOfIndices::with_index_bound(16));
+/// test_priority_queue_deckey(BinaryHeapWithMap::default());
+/// test_priority_queue_deckey(BinaryHeapWithMap::with_capacity(16));
+/// test_priority_queue_deckey(QuarternaryHeapWithMap::default());
+/// test_priority_queue_deckey(QuarternaryHeapWithMap::with_capacity(16));
+/// test_priority_queue_deckey(QuarternaryHeapWithMap::default());
+/// test_priority_queue_deckey(QuarternaryHeapWithMap::with_capacity(16));
 /// ```
-#[derive(Clone, Debug)]
-pub struct DaryHeapOfIndices<N, K, const D: usize = 2>
+#[derive(Debug, Clone)]
+pub struct DaryHeapWithMap<N, K, const D: usize = 2>
 where
-    N: HasIndex,
+    N: Index,
     K: PartialOrd + Clone,
 {
-    heap: Heap<N, K, HeapPositionsHasIndex<N>, D>,
+    heap: Heap<N, K, HeapPositionsMap<N>, D>,
 }
 
-impl<N, K, const D: usize> DaryHeapOfIndices<N, K, D>
+impl<N, K, const D: usize> Default for DaryHeapWithMap<N, K, D>
 where
-    N: HasIndex,
+    N: Index,
     K: PartialOrd + Clone,
 {
-    /// As explained in [`DaryHeapOfIndices`],
-    /// this heap is useful when the nodes come from a closed set with a known size.
-    /// Therefore, the heap has a strict exclusive upper bound on the index of a node which can enter the heap,
-    /// defined by the argument `with_index_bound`.
+    fn default() -> Self {
+        Self {
+            heap: Heap::new(None, HeapPositionsMap::default()),
+        }
+    }
+}
+impl<N, K, const D: usize> DaryHeapWithMap<N, K, D>
+where
+    N: Index,
+    K: PartialOrd + Clone,
+{
+    /// Creates a new empty d-ary heap.
     ///
-    /// The closed set of indices which can enter the heap is [0, 1, ..., `index_bound`).
-    ///
-    /// The upper bound on the indices of a `DaryHeapOfIndices` can be obtained by the `index_bound` method.
-    ///
-    /// # Examples
+    ///  # Examples
     ///
     /// ```
     /// use orx_priority_queue::*;
     ///
-    /// // set of possible nodes which can enter the heap is closed and has 16 elements
-    /// let mut pq = BinaryHeapOfIndices::with_index_bound(16);
+    /// let mut heap = BinaryHeapWithMap::new();
     ///
-    /// assert_eq!(16, pq.index_bound());
+    /// heap.push('a', 4);
+    /// heap.push('b', 42);
     ///
-    /// // 8-th node enters the queue with key of 100.0
-    /// pq.push(7usize, 100.0);
-    ///
-    /// // third node enters
-    /// pq.push(2, 42.0);
-    ///
-    /// // the following line would've panicked since there exist no node with index 16 in the closed set [0, 1, ..., 15]
-    /// // pq.push(16, 7.0);
+    /// assert_eq!(Some('a'), heap.pop_node());
+    /// assert_eq!(Some('b'), heap.pop_node());
+    /// assert!(heap.is_empty());
     /// ```
-    pub fn with_index_bound(index_bound: usize) -> Self {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Creates a new d-ary heap with the given initial `capacity` on the number of nodes to simultaneously exist on the heap.
+    pub fn with_capacity(capacity: usize) -> Self {
         Self {
-            heap: Heap::new(None, HeapPositionsHasIndex::with_index_bound(index_bound)),
+            heap: Heap::new(Some(capacity), HeapPositionsMap::with_capacity(capacity)),
         }
     }
-
-    /// Cardinality of the closed set which the nodes are sampled from.
-    ///
-    /// # Panics
-    ///
-    /// Panics if a node with an index greater than or equal to the `index_bound` is pushed to the queue.
-    pub fn index_bound(&self) -> usize {
-        self.heap.positions().index_bound()
-    }
-
     /// Returns the 'd' of the d-ary heap.
     /// In other words, it represents the maximum number of children that each node on the heap can have.
     pub const fn d() -> usize {
@@ -222,9 +202,9 @@ where
     }
 }
 
-impl<N, K, const D: usize> PriorityQueue<N, K> for DaryHeapOfIndices<N, K, D>
+impl<N, K, const D: usize> PriorityQueue<N, K> for DaryHeapWithMap<N, K, D>
 where
-    N: HasIndex,
+    N: Index,
     K: PartialOrd + Clone,
 {
     type NodeKey<'a> = &'a (N, K) where Self: 'a, N: 'a, K: 'a;
@@ -234,61 +214,74 @@ where
     fn len(&self) -> usize {
         self.heap.len()
     }
+
     #[inline(always)]
     fn capacity(&self) -> usize {
         self.heap.capacity()
     }
+
     fn peek(&self) -> Option<&(N, K)> {
         self.heap.peek()
     }
+
     fn clear(&mut self) {
         self.heap.clear()
     }
+
     #[inline(always)]
     fn pop(&mut self) -> Option<(N, K)> {
         self.heap.pop()
     }
+
     #[inline(always)]
     fn pop_node(&mut self) -> Option<N> {
         self.heap.pop_node()
     }
+
     #[inline(always)]
     fn pop_key(&mut self) -> Option<K> {
         self.heap.pop_key()
     }
+
     #[inline(always)]
     fn push(&mut self, node: N, key: K) {
         self.heap.push(node, key)
     }
+
     #[inline(always)]
     fn push_then_pop(&mut self, node: N, key: K) -> (N, K) {
         self.heap.push_then_pop(node, key)
     }
+
     fn iter(&self) -> Self::Iter<'_> {
         self.as_slice().iter()
     }
 }
-impl<N, K, const D: usize> PriorityQueueDecKey<N, K> for DaryHeapOfIndices<N, K, D>
+impl<N, K, const D: usize> PriorityQueueDecKey<N, K> for DaryHeapWithMap<N, K, D>
 where
-    N: HasIndex,
+    N: Index,
     K: PartialOrd + Clone,
 {
     #[inline(always)]
     fn contains(&self, node: &N) -> bool {
         self.heap.contains(node)
     }
+
     #[inline(always)]
     fn key_of(&self, node: &N) -> Option<K> {
         self.heap.key_of(node)
     }
+
     #[inline(always)]
     fn decrease_key(&mut self, node: &N, decreased_key: K) {
         self.heap.decrease_key(node, decreased_key)
     }
+
     #[inline(always)]
     fn update_key(&mut self, node: &N, new_key: K) -> ResUpdateKey {
         self.heap.update_key(node, new_key)
     }
+
     #[inline(always)]
     fn remove(&mut self, node: &N) -> K {
         self.heap.remove(node)
