+++ src/has_index.rs
