+++ src/impl_queues/priority_queue.rs
@@ -14,38 +14,47 @@ where
     fn len(&self) -> usize {
         priority_queue::PriorityQueue::len(self)
     }
+
     fn capacity(&self) -> usize {
         priority_queue::PriorityQueue::capacity(self)
     }
+
     #[inline(always)]
     fn peek(&self) -> Option<Self::NodeKey<'_>> {
         priority_queue::PriorityQueue::peek(self)
     }
+
     #[inline(always)]
     fn clear(&mut self) {
         priority_queue::PriorityQueue::clear(self)
     }
+
     #[inline(always)]
     fn pop(&mut self) -> Option<(N, K)> {
         priority_queue::PriorityQueue::pop(self)
     }
+
     #[inline(always)]
     fn pop_node(&mut self) -> Option<N> {
         priority_queue::PriorityQueue::pop(self).map(|x| x.0)
     }
+
     #[inline(always)]
     fn pop_key(&mut self) -> Option<K> {
         priority_queue::PriorityQueue::pop(self).map(|x| x.1)
     }
+
     #[inline(always)]
     fn push(&mut self, node: N, key: K) {
         priority_queue::PriorityQueue::push(self, node, key);
     }
+
     #[inline(always)]
     fn push_then_pop(&mut self, node: N, key: K) -> (N, K) {
         priority_queue::PriorityQueue::push(self, node, key);
         priority_queue::PriorityQueue::pop(self).expect("queue is not empty")
     }
+
     fn iter(&self) -> Self::Iter<'_> {
         priority_queue::PriorityQueue::iter(self)
     }
@@ -60,10 +69,12 @@ where
     fn contains(&self, node: &N) -> bool {
         priority_queue::PriorityQueue::get(self, node).is_some()
     }
+
     #[inline(always)]
     fn key_of(&self, node: &N) -> Option<K> {
         priority_queue::PriorityQueue::get(self, node).map(|x| x.1.clone())
     }
+
     fn decrease_key(&mut self, node: &N, decreased_key: K) {
         let old_key =
             priority_queue::PriorityQueue::change_priority(self, node, decreased_key.clone())
@@ -75,6 +86,7 @@ where
         })
         .expect("Failed to decrease the key of the node, received a greater key");
     }
+
     fn update_key(&mut self, node: &N, new_key: K) -> ResUpdateKey {
         let old_key = priority_queue::PriorityQueue::change_priority(self, node, new_key.clone())
             .expect("Failed to update key of the node, it is not present in the queue");
@@ -84,6 +96,7 @@ where
             ResUpdateKey::Increased
         }
     }
+
     #[inline(always)]
     fn remove(&mut self, node: &N) -> K {
         priority_queue::PriorityQueue::remove(self, node)
