+++ tests/daryheap_with_map.rs
