+++ tests/daryheap_with_map_2.rs
