+++ tests/dijkstra.rs
